requirements.txt
datasets==4.2.0
matplotlib==3.10.7
numpy==2.3.4
Pillow==12.0.0
torch==2.6.0
tqdm==4.67.1
transformers==4.57.1
